requirements.txt
langchain>=0.2.10
langchain-community>=0.2.10
langchain-core>=0.2.10
langchain-openai>=0.1.8
langchain-text-splitters>=0.2.2
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.2
chromadb>=0.5.5
pypdf>=4.2.0
sentence-transformers>=3.0.1
python-dotenv>=1.0.1
openai>=1.40.0
tiktoken>=0.7.0
python-pptx>=0.6.23
pymupdf>=1.24.9
pillow>=10.4.0
pytesseract>=0.3.13
